+++ test/unit/dal/mutations/business/update-business-name.test.js
@@ -1,10 +1,10 @@
 import { parse } from 'graphql'
 import { describe, test, expect } from 'vitest'
-import { updateBusinessAddressMutation } from '../../../../src/dal/mutations/business/update-business-address.js'
+import { updateBusinessNameMutation } from '../../../../../src/dal/mutations/business/update-business-name.js'
 
-describe('When the updateBusinessAddressMutation is parsed', () => {
+describe('When the updateBusinessNameMutation is parsed', () => {
   test('it is valid GraphQL syntax', () => {
-    expect(() => parse(updateBusinessAddressMutation)).not.toThrow()
+    expect(() => parse(updateBusinessNameMutation)).not.toThrow()
   })
 
   test('it contains the Mutation operation and the correct variable', () => {
@@ -12,12 +12,12 @@ describe('When the updateBusinessAddressMutation is parsed', () => {
     // this converts the string into a structural representation of the mutation.
     // This allows it be inspected and validated.
 
-    const ast = parse(updateBusinessAddressMutation)
+    const ast = parse(updateBusinessNameMutation)
     const operation = ast.definitions[0]
     expect(operation.name.value).toBe('Mutation')
 
     const variable = operation.variableDefinitions[0]
     expect(variable.variable.name.value).toBe('input')
-    expect(variable.type.type.name.value).toBe('UpdateBusinessAddressInput')
+    expect(variable.type.type.name.value).toBe('UpdateBusinessNameInput')
   })
 })
